Implement setup logic. API can't be active unless it was successfully added to the database
