+++ MultiplayerWebFramework/HostServicesAPI/Program.cs
@@ -25,6 +25,6 @@ namespace HostServicesAPI
                 .ConfigureWebHostDefaults(webBuilder =>
                 {
                     webBuilder.UseStartup<Startup>();
-                }).ConfigureServices(services => services.AddHostedService<LifetimeEventsHostedService>());     // Add our IHostedService
+                }).ConfigureServices(services => services.AddHostedService<SetupTeardownHostedService>());     // Add our IHostedService
     }
 }
